Initial Commit
This project isolates an issue where Stage3D and StageVideo interact
to produce an unwanted result that is hard to work around.
When a NetStream instance is disposed (on an iPad or other device)
display objects on the main Flash stage (2D) lose their transparency
and become opaque/black.  This obscures Stage3D below.